+++ src/main/java/org/sopt/user/infrastructure/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package org.sopt.user.repository;
+package org.sopt.user.infrastructure.repository;
 
 import org.sopt.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
