+++ src/main/java/org/sopt/user/service/UserService.java
@@ -2,7 +2,7 @@ package org.sopt.user.service;
 
 import org.sopt.user.domain.User;
 import org.sopt.user.presentation.dto.request.CreateUserRequest;
-import org.sopt.user.repository.UserRepository;
+import org.sopt.user.infrastructure.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 @Service
